The changes made to the `go.mod` and `go.sum` files indicate that the following packages have been updated:
1. `github.com/urfave/cli/v2`: The version of this package has been updated from `v0.0.0-20161208181325-20d25e280405` to a newer version.
2. `golang.org/x/net`: The version of this package has been updated from `v0.34.0` to a newer version, specifically `v0.34.1`.
3. `golang.org/x/text`: The version of this package has been updated from `v0.21.0` to a newer version, specifically `v0.21.1`.

Additionally, the `env_test.go` file has been modified to remove the line that was trimming the `--` prefix from the `headersRaw` variable.

The changes in the `form.go` file are related to the parsing of form data and headers. The lines that were trimming the `--` prefix from the `headersRaw` variable have been replaced with calls to `strings.TrimPrefix`, which is a more efficient way to remove the prefix.

The changes in the `request_test.go` file do not appear to be related to any specific package updates, but rather seem to be related to testing the `SplitRequests` function.